added PresetColorPack and started adding colors again
